+++ lib/hcloud/concerns/updatable.rb
@@ -13,6 +13,8 @@ module HCloud
         assign_attributes client
           .put("/#{resource_name.pluralize}/#{id}", updatable_params)
           .fetch(resource_name.to_sym)
+
+        self
       end
 
       def updatable_attributes