+++ spec/hcloud/concerns/creatable_spec.rb
@@ -15,10 +15,19 @@ RSpec.describe HCloud::Creatable do
       stub_request(:post, "https://api.hetzner.cloud/v1/examples")
         .with(body: { name: "my_resource", description: "my_description", sibling: { type: "sister", child: { id: 1 } }, child: "name1", children: [1, nil] })
         .to_return(body: { example: resource.attributes.merge(id: 1, created: 1.second.ago) }.to_json)
+
       resource.create
 
       expect(resource.id).to eq 1
       expect(resource).to be_created
+    end
+
+    it "returns the resource" do
+      stub_request(:post, "https://api.hetzner.cloud/v1/examples")
+        .with(body: { name: "my_resource", description: "my_description", sibling: { type: "sister", child: { id: 1 } }, child: "name1", children: [1, nil] })
+        .to_return(body: { example: resource.attributes.merge(id: 1, created: 1.second.ago) }.to_json)
+
+      expect(resource.create).to be_a described_class
     end
   end