+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 - Allow calling `#update` with additional attributes to be updated
+- Return the resource when calling `#create`, `#update` or `#delete`
 - Add additional primary IP mismatch errors:
   - `PrimaryIPAssigned`
   - `PrimaryIPDatacenterMismatch`